pyproject.toml
[tool.poetry]
name = "lector-facturas"
version = "0.1.0"
description = "Aplicación para leer facturas usando Ollama y Gemma3 12B"
authors = ["Felipe Mancini <felipe@asimov.cl>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
python-ollama = "^0.2.4"
pdf2image = "^1.17.0"
pytesseract = "^0.3.13"
Pillow = "^11.2.0"
streamlit = "^1.44.1"
python-dotenv = "^1.1.0"
requests = "^2.31.0"
ollama = "^0.4.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
